requirements.txt
# LexAI Backend — Python Dependencies
# Install: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── LLM Client ───────────────────────────────────────
anthropic>=0.28.0

# ── Config & Env ─────────────────────────────────────
python-dotenv==1.0.1
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Authentication ────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── Rate Limiting ─────────────────────────────────────
slowapi==0.1.9

# ── Database (SQLite via SQLAlchemy) ──────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0

# ── PDF Export ────────────────────────────────────────
reportlab==4.2.0

# ── Logging ───────────────────────────────────────────
loguru==0.7.2

# ── Testing ───────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# ── RAG (Optional — real semantic search) ─────────────
# sentence-transformers==3.0.1
# faiss-cpu==1.8.0
# numpy>=1.26.0
